feat: initState
