+++ src/constants.ts
@@ -1,0 +1,3 @@
+import { Size } from "./interfaces";
+
+export const sizes: Size[] = [4, 6, 8];
